Clean up layout
size related functions were moved into their own module.
frames got its own directory.
