+++ lib/mux/src/lib.rs
@@ -1,54 +1,12 @@
 extern crate byteorder;
 
-use std::io;
-use std::io::{Cursor, Read, Write};
-
 use std::time::Duration;
 
-use byteorder::{ReadBytesExt, WriteBytesExt, BigEndian};
+pub mod types;
+pub mod frames;
 
 pub type Headers = Vec<(u8, Vec<u8>)>;
 pub type Contexts = Vec<(Vec<u8>, Vec<u8>)>;
-
-pub mod types {
-    pub const TREQ: i8 = 1;
-    pub const RREQ: i8 = -1;
-
-    pub const TDISPATCH: i8 = 2;
-    pub const RDISPATCH: i8 = -2;
-
-    pub const TINIT: i8 = 68;
-    pub const RINIT: i8 = -68;
-
-    pub const TDRAIN: i8 = 64;
-    pub const RDRAIN: i8 = -64;
-
-    pub const TPING: i8 = 65;
-    pub const RPING: i8 = -65;
-
-    pub const TDISCARDED: i8 = 66;
-    pub const TLEASE: i8 = 67;
-
-    pub const RERR: i8 = -128;
-}
-
-// extract a value from the byteorder::Result
-macro_rules! tryb {
-    ($e:expr) => (
-        match $e {
-            Ok(r) => r,
-            Err(byteorder::Error::UnexpectedEOF) => {
-                return Err(io::Error::new(
-                    io::ErrorKind::UnexpectedEof,
-                    "End of input"
-                ));
-            }
-            Err(byteorder::Error::Io(err)) => {
-                return Err(err);
-            }
-        }
-    )
-}
 
 #[derive(Debug, PartialEq, Eq)]
 pub struct MuxPacket {
@@ -97,18 +55,6 @@ pub struct Treq {
     pub body: Vec<u8>,
 }
 
-impl Treq {
-    pub fn frame_size(&self) -> usize {
-        let mut size = 1; // header count
-        for &(_, ref v) in &self.headers {
-            size += 2; // key and value lengths
-            size += v.len();
-        }
-
-        size + self.body.len()
-    }
-}
-
 #[derive(PartialEq, Eq, Debug)]
 pub struct Tdispatch {
     pub contexts: Contexts,
@@ -128,17 +74,6 @@ pub enum Rmsg {
     Ok(Vec<u8>),
     Error(String),
     Nack(String),
-}
-
-impl Rmsg {
-    #[inline]
-    pub fn msg_size(&self) -> usize {
-        match self {
-            &Rmsg::Ok(ref b) => b.len(),
-            &Rmsg::Error(ref m) => m.as_bytes().len(),
-            &Rmsg::Nack(ref m) => m.as_bytes().len(),
-        }
-    }
 }
 
 #[derive(PartialEq, Eq, Debug)]
@@ -164,18 +99,6 @@ impl Tag {
     }
 }
 
-impl Init {
-    pub fn frame_size(&self) -> usize {
-        let mut size = 2; // version
-
-        for &(ref k, ref v) in &self.headers {
-            // each value preceeded by its len (i32)
-            size += 8 + k.len() + v.len();
-        }
-        size
-    }
-}
-
 impl Dtab {
     #[inline]
     pub fn new() -> Dtab {
@@ -194,23 +117,6 @@ impl Dtab {
 }
 
 impl MessageFrame {
-    pub fn frame_size(&self) -> usize {
-        match self {
-            &MessageFrame::Treq(ref f) => f.frame_size(),
-            &MessageFrame::Rreq(ref f) => 1 + f.msg_size(),
-            &MessageFrame::Tdispatch(ref f) => f.frame_size(),
-            &MessageFrame::Rdispatch(ref f) => f.frame_size(),
-            &MessageFrame::Tinit(ref f) => f.frame_size(),
-            &MessageFrame::Rinit(ref f) => f.frame_size(),
-            &MessageFrame::Tdrain => 0,
-            &MessageFrame::Rdrain => 0,
-            &MessageFrame::Tping => 0,
-            &MessageFrame::Rping => 0,
-            &MessageFrame::Tlease(_) => 9,
-            &MessageFrame::Rerr(ref msg) => msg.as_bytes().len(),
-        }
-    }
-
     pub fn frame_id(&self) -> i8 {
         match self {
             &MessageFrame::Treq(_) => types::TREQ,
@@ -229,42 +135,7 @@ impl MessageFrame {
     }
 }
 
-#[inline]
-fn context_size(contexts: &Contexts) -> usize {
-    let mut size = 2; // context size
-
-    for &(ref k, ref v) in contexts {
-        size += 4; // two lengths
-        size += k.len();
-        size += v.len();
-    }
-    size
-}
-
-#[inline]
-fn dtab_size(table: &Dtab) -> usize {
-    let mut size = 2; // context size
-
-    for &(ref k, ref v) in &table.entries {
-        size += 4; // the two lengths
-        size += k.as_bytes().len();
-        size += v.as_bytes().len();
-    }
-
-    size
-}
-
 impl Tdispatch {
-    fn frame_size(&self) -> usize {
-        let mut size = 2 + // dest size
-                       context_size(&self.contexts) +
-                       dtab_size(&self.dtab);
-
-        size += self.dest.as_bytes().len();
-        size += self.body.len();
-        size
-    }
-
     pub fn basic_(dest: String, body: Vec<u8>) -> Tdispatch {
         Tdispatch {
             contexts: Vec::new(),
@@ -279,120 +150,6 @@ impl Tdispatch {
     }
 }
 
-impl Rdispatch {
-    fn frame_size(&self) -> usize {
-        1 + context_size(&self.contexts) + match &self.msg {
-            &Rmsg::Ok(ref body) => body.len(),
-            &Rmsg::Error(ref msg) => msg.as_bytes().len(),
-            &Rmsg::Nack(ref msg) => msg.as_bytes().len(),
-        }
-    }
-}
-
-// write the message to the Write
-pub fn encode_message(buffer: &mut Write, msg: &Message) -> io::Result<()> {
-    // the size is the buffer size + the header (id + tag)
-    tryb!(buffer.write_i32::<BigEndian>(msg.frame.frame_size() as i32 + 4));
-    tryb!(buffer.write_i8(msg.frame.frame_id()));
-    try!(frames::encode_tag(buffer, &msg.tag));
-
-    encode_frame(buffer, &msg.frame)
-}
-
-pub fn encode_frame(buffer: &mut Write, frame: &MessageFrame) -> io::Result<()> {
-    match frame {
-        &MessageFrame::Treq(ref f) => frames::encode_treq(buffer, f),
-        &MessageFrame::Rreq(ref f) => frames::encode_rreq(buffer, f),
-        &MessageFrame::Tdispatch(ref f) => frames::encode_tdispatch(buffer, f),
-        &MessageFrame::Rdispatch(ref f) => frames::encode_rdispatch(buffer, f),
-        &MessageFrame::Tinit(ref f) => frames::encode_init(buffer, f),
-        &MessageFrame::Rinit(ref f) => frames::encode_init(buffer, f),
-        // the following are empty messages
-        &MessageFrame::Tping => Ok(()),
-        &MessageFrame::Rping => Ok(()),
-        &MessageFrame::Tdrain => Ok(()),
-        &MessageFrame::Rdrain => Ok(()),
-        &MessageFrame::Tlease(ref d) => {
-            let millis = d.as_secs()*1000 + (((d.subsec_nanos() as f64)/1e6) as u64);
-            tryb!(buffer.write_u8(0));
-            tryb!(buffer.write_u64::<BigEndian>(millis));
-            Ok(())
-        }
-        &MessageFrame::Rerr(ref msg) => {
-            buffer.write_all(msg.as_bytes())
-        }
-    }
-}
-
-// Decodes `data` into a frame if type `tpe`
-pub fn decode_frame(tpe: i8, data: &[u8]) -> io::Result<MessageFrame> {
-    Ok(match tpe {
-        types::TREQ => MessageFrame::Treq(try!(frames::decode_treq(data))),
-        types::RREQ => MessageFrame::Rreq(try!(frames::decode_rreq(data))),
-        types::TDISPATCH => MessageFrame::Tdispatch(try!(frames::decode_tdispatch(data))),
-        types::RDISPATCH => MessageFrame::Rdispatch(try!(frames::decode_rdispatch(data))),
-        types::TINIT => MessageFrame::Tinit(try!(frames::decode_init(data))),
-        types::RINIT => MessageFrame::Rinit(try!(frames::decode_init(data))),
-        types::TDRAIN => MessageFrame::Tdrain,
-        types::RDRAIN => MessageFrame::Rdrain,
-        types::TPING => MessageFrame::Tping,
-        types::RPING => MessageFrame::Rping,
-        types::TLEASE => {
-            let mut buffer = Cursor::new(data);
-            let _ = try!(buffer.read_u8());
-            let ticks = try!(buffer.read_u64::<BigEndian>());
-            MessageFrame::Tlease(Duration::from_millis(ticks))
-        }
-        types::RERR => MessageFrame::Rerr(try!(frames::decode_rerr(data))),
-        other => {
-            return Err(
-                io::Error::new(io::ErrorKind::InvalidInput,
-                    format!("Invalid frame type: {}", other))
-                );
-        }
-    })
-}
-
-// Read an entire frame buffer
-pub fn read_frame(input: &mut Read) -> io::Result<MuxPacket> {
-    let size = {
-        let size = tryb!(input.read_i32::<BigEndian>());
-        if size < 4 {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData, "Invalid mux frame size"
-            ));
-        }
-        size as usize
-    };
-
-    let tpe = tryb!(input.read_i8());
-    let tag = try!(frames::decode_tag(input));
-
-    let mut buf = vec![0;size-4];
-    try!(input.read_exact(&mut buf));
-    Ok(MuxPacket {
-        tpe: tpe,
-        tag: tag,
-        buffer: buf,
-    })
-}
-
-// This is a synchronous function that will read a whole message from the `Read`
-pub fn read_message(input: &mut Read) -> io::Result<Message> {
-    let packet = try!(read_frame(input));
-    decode_message(&packet)
-}
-
-// expects a SharedReadBuffer of the whole mux frame
-pub fn decode_message(input: &MuxPacket) -> io::Result<Message> {
-    let frame = try!(decode_frame(input.tpe, &input.buffer));
-    Ok(Message {
-        tag: input.tag.clone(),
-        frame: frame,
-    })
-}
-
-pub mod frames;
 
 #[cfg(test)]
 mod tests;
